move util functions to utils.py
